+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added clone action to energy storage container in myems-api
 - added export action to energy storage container in myems-api
+- added actions for protocol in myems-api
 ### Changed
 - updated tariff price to keep 5 decimal places in database and myems-admin 
 ### Fixed